Move NodeLocation to parser module
