+++ crates/apollo-compiler/src/ast/from_cst.rs
@@ -1,9 +1,9 @@
 use crate::ast;
 use crate::ast::Document;
 use crate::parser::FileId;
+use crate::parser::NodeLocation;
 use crate::parser::SourceMap;
 use crate::Node;
-use crate::NodeLocation;
 use apollo_parser::cst;
 use apollo_parser::cst::CstNode;
 use apollo_parser::SyntaxNode;
