+++ crates/apollo-compiler/src/lib.rs
@@ -18,6 +18,5 @@ pub use self::executable::ExecutableDocument;
 pub use self::name::InvalidNameError;
 pub use self::name::Name;
 pub use self::node::Node;
-pub use self::node::NodeLocation;
 pub use self::parser::parse_mixed_validate;
 pub use self::schema::Schema;
