+++ crates/apollo-compiler/src/name.rs
@@ -2,11 +2,11 @@ use crate::diagnostic::CliReport;
 use crate::diagnostic::ToCliReport;
 use crate::execution::GraphQLLocation;
 use crate::parser::FileId;
+use crate::parser::NodeLocation;
 use crate::parser::SourceMap;
 use crate::parser::TaggedFileId;
 use crate::schema::ComponentName;
 use crate::schema::ComponentOrigin;
-use crate::NodeLocation;
 use rowan::TextRange;
 use std::fmt;
 use std::marker::PhantomData;
