+++ crates/apollo-compiler/src/schema/mod.rs
@@ -4,9 +4,9 @@ use crate::ast;
 use crate::collections::IndexMap;
 use crate::collections::IndexSet;
 use crate::parser::FileId;
+use crate::parser::NodeLocation;
 use crate::parser::Parser;
 use crate::Node;
-use crate::NodeLocation;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::path::Path;
