+++ crates/apollo-compiler/src/validation/mod.rs
@@ -34,12 +34,12 @@ use crate::executable::ConflictingFieldType;
 use crate::executable::VariableDefinition;
 use crate::execution::GraphQLError;
 use crate::execution::Response;
+use crate::parser::NodeLocation;
 use crate::parser::SourceMap;
 use crate::schema::BuildError as SchemaBuildError;
 use crate::schema::Implementers;
 use crate::Name;
 use crate::Node;
-use crate::NodeLocation;
 use std::collections::HashMap;
 use std::fmt;
 use std::sync::Arc;
